refactor google connectivity and tools locations
